requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
lxml==4.3.2
mysql-connector-python-rf==2.2.2
progressbar==2.5
python-utils==2.3.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9
SQLAlchemy==1.3.1
urllib3==1.24.1
